a7: fix up the page linking
