+++ public/tuiter/explore-old/explore-settings.html
@@ -1,0 +1,9 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+   <meta charset="UTF-8">
+   <title>Explore Settings</title>
+</head>
+<body>
+</body>
+</html>
