+++ public/tuiter/explore-old/index.css
@@ -1,0 +1,9 @@
+.fa-magnifying-glass {
+    position: absolute;
+    padding-left: 10px;
+}
+
+#search-text {
+    box-sizing: border-box;
+    padding-left: 35px;
+}
