+++ public/tuiter/explore-old/index.html
@@ -1,0 +1,18 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.1.0/dist/css/bootstrap.min.css" rel="stylesheet"/>
+    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/bootswatch/5.2.3/cyborg/bootstrap.min.css" integrity="sha512-6JM8AUNiR0PhS+P/orbIggnQlkVEG7fUoO6ve7XWNGGGL0mOfLPSYuBo4/31Ig5KiC1T9Ep5yw8R12E2qaI2hQ==" crossorigin="anonymous" referrerpolicy="no-referrer" rel="stylesheet"/>
+    <link href="index.css" rel="stylesheet"/>
+    <script src="https://kit.fontawesome.com/017c1efe25.js" crossorigin="anonymous"></script>
+    <script src="https://code.jquery.com/jquery-3.6.1.min.js"
+           integrity="sha256-o88AwQnZB+VDvE9tvIXrMQaPlFFSUTR+nldQm1LuPXQ="
+           crossorigin="anonymous"></script>
+   <meta charset="UTF-8">
+   <title>Explore screen</title>
+</head>
+<body>
+    <div class="container" id="wd-explore"></div>
+    <script src="index.js" type="module"></script>
+</body>
+</html>
