--- public/tuiter/explore/explore-settings.html
@@ -1,9 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-   <meta charset="UTF-8">
-   <title>Explore Settings</title>
-</head>
-<body>
-</body>
-</html>
