--- public/tuiter/explore/index.css
@@ -1,9 +1,0 @@
-.fa-magnifying-glass {
-    position: absolute;
-    padding-left: 10px;
-}
-
-#search-text {
-    box-sizing: border-box;
-    padding-left: 35px;
-}
