+++ public/tuiter/navigation.html
@@ -3,7 +3,7 @@
     <a href="home.html"> Home </a>
   </li>
   <li>
-    <a href="explore/index.html"> Explore </a>
+    <a href="explore-old/index.html"> Explore </a>
   </li>
   <li>
     <a href="notifications.html"> Notifications </a>
