+++ src/tuiter/explore/index.js
@@ -2,7 +2,7 @@ import React from "react";
 import PostSummaryList from "../post-summary-list";
 import "./index.css";
 
-const ExploreComponent = () => {
+const Explore = () => {
  return(
    <>
      <div className="row">
@@ -38,4 +38,4 @@ const ExploreComponent = () => {
  );
 };
 
-export default ExploreComponent;
+export default Explore;
