+++ src/tuiter/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import { Provider } from "react-redux";
+import {useParams} from "react-router";
 import NavigationSidebar from "./navigation-sidebar";
 import WhoToFollowList from "./who-to-follow-list";
 import whoReducer
@@ -12,11 +13,12 @@ const store = configureStore(
   {reducer: {who: whoReducer, tuits: tuitsReducer}});
 
 function Tuiter() {
+  const { path } = useParams();
   return (
     <Provider store={store}>
       <div className="row mt-2">
         <div className="col-2 col-md-2 col-lg-1 col-xl-2">
-          <NavigationSidebar active="explore"/>
+          <NavigationSidebar active={path || 'home'}/>
         </div>
         <div className="col-10 col-md-10 col-lg-7 col-xl-6"
               style={{"position": "relative"}}>
